Removed redundant test file and moved tests over, also modified database to include link for photo and wiki link
