+++ pokinaturalist/tests.py
@@ -1,6 +1,8 @@
 from django.test import TestCase, Client
 from django.core import exceptions
 from django.http import HttpRequest
+from django.db import IntegrityError
+from .models import User, Move, Creature, Item
 
 from . import views
 from ipware import get_client_ip
@@ -31,3 +33,90 @@ class PokinaturalistAppTestCase(TestCase):
         request = HttpRequest()
         request.META = {}
         self.assertIsNone(views.get_user_ip_addr(request))
+
+
+# Database  tests(for local database)
+class DatabaseTests(TestCase):
+    @classmethod
+    def setUp(cls):
+        test = User(token_id="893", username="Ash Ketchum", points=25)
+        test.save()
+
+    def test_user_creation(self):
+        test2 = User(token_id="6484", username="Gary Oak", points=50)
+        test2.save()
+        self.assertEqual(User.objects.get(username="Gary Oak"), test2)
+
+    # add a duplicate user test just to be sure of unique status of username
+    def test_duplicate_username(self):
+        try:
+            test2 = User(token_id="8064", username="Ash Ketchum", points=12)
+            test2.save()
+        except IntegrityError:
+            return True
+        return False
+    def test_move_creation(self):
+        test2 = Move(move_name="Thundershock", move_strength=50)
+        self.assertEqual(Move.objects.get(), test2)
+    # add a duplicate move test just to be sure of unique status of username
+    def test_duplicate_move(self):
+        try:
+            test2 = Move(move_name="Thundershock", move_strength=30)
+            test3 = Move(move_name="Thundershock", move_strength=50)
+            test2.save()
+            test3.save()
+        except IntegrityError:
+            return True
+        return False
+    def test_creature_creation(self):
+        test2 = User.objects.get(username="Ash Ketchum")
+        test3 = Creature(owner_id=test2, nickname="Pikachu",
+                         observation_icon="Pokemon num 25", observation_wiki="Pokemon num 25", strength=1337)
+        test4 = Move(move_name="Thunderbolt", move_strength=42)
+        test3.save()
+        test4.save()
+        test3.moveset.add(test4)
+        self.assertEqual(Creature.objects.get(), test3)
+
+    def test_item_creation(self):
+        test2 = User.objects.get(username="Ash Ketchum")
+        test3 = Item(item_type=1, item_strength=50)
+        test3.owner_id = test2
+        test3.save()
+        self.assertEqual(Item.objects.get(), test3)
+
+    #user to creatures
+    def test_multiple_creature(self):
+        # this will test interaction between user and creature by attaching multiple pokemon
+        # to the setup of THE POKEMON MASTER(Ash)
+        test2 = User.objects.get(username="Ash Ketchum")
+        test3 = Creature(owner_id=test2, nickname="Pikachu",
+                         observation_icon="Defibrillator", observation_wiki="Pokemon num 25", strength=1337)
+        test4 = Creature(owner_id=test2, nickname="Charizard",
+                         observation_icon="Blaze it", observation_wiki="Pokemon num 6", strength=420)
+        test3.save()
+        test4.save()
+        self.assertEqual(Creature.objects.filter(owner_id=test2).count(), 2)
+    #creature to moves
+    def test_multiple_moves(self):
+        test2 = User.objects.get(username="Ash Ketchum")
+        test3 = Creature(owner_id=test2, nickname="Pikachu", observation_base="Pokemon num 25", strength=1337)
+        test4 = Move(move_name="Thunderbolt", move_strength=42)
+        test5 = Move(move_name="Volt Tackle", move_strength=69)
+        test3.save()
+        test4.save()
+        test5.save()
+        test3.moveset.add(test4)
+        test3.moveset.add(test5)
+        self.assertEqual(Creature.objects.get(nickname="Pikachu").moveset.count(), 2)
+
+    #user to items
+    def test_multiple_items(self):
+        # let's assume ash has 2 potions
+        test2 = User.objects.get(username="Ash Ketchum")
+        test3 = Item(owner_id=test2, item_type=1, item_strength=50)
+        test4 = Item(owner_id=test2, item_type=1, item_strength=50)
+        test3.save()
+        test4.save()
+        self.assertEqual(Item.objects.filter(owner_id=test2).count(), 2)
+
